refactor(api): move email-verification serializer
to src
