+++ api/src/identity-access-management/infrastructure/serializers/jsonapi/email-verification.serializer.js
@@ -11,4 +11,5 @@ const deserialize = function (payload) {
   });
 };
 
-export { deserialize };
+const emailVerificationSerializer = { deserialize };
+export { emailVerificationSerializer };
